+++ qa.py
@@ -1,0 +1,86 @@
+import re
+from typing import List, Optional
+
+import requests
+from fastapi import HTTPException
+
+from config import OLLAMA_KEEP_ALIVE, OLLAMA_URL, resolve_model
+
+
+def dehedge(text: str) -> str:
+    patterns = [
+        r'(?i)^\s*(according to|based on|from)\s+(the\s+)?(provided|given)\s+(context|information|documents)\s*[:,\-]*\s*',
+        r'(?i)^\s*(according to|based on)\s+(the\s+)?document(s)?\s*[:,\-]*\s*',
+    ]
+    for p in patterns:
+        text = re.sub(p, "", text, count=1)
+    return text.strip()
+
+
+def ask_with_context(question: str, hits: List[dict], chat_history: Optional[List[dict]] = None, model: Optional[str] = None) -> str:
+    ql = (question or "").lower()
+
+    meta_triggers = [
+        "your name", "what is your name", "what's your name",
+        "who are you", "who is adam", "what does adam stand for",
+        "what are you", "introduce yourself"
+    ]
+    if any(t in ql for t in meta_triggers):
+        return "I am Adam - the Amentum Document and Assistance Model (ADAM)."
+
+    context = "\n\n".join([f"[{i+1}] {h['text']}" for i, h in enumerate(hits)])
+
+    sys_prompt = (
+        "You are Adam — the Amentum Document and Assistance Model (ADAM). "
+        "Answer directly and succinctly. Do not start with phrases like "
+        "'According to the provided context'. Use ONLY the provided context for factual claims and insert "
+        "inline bracket citations like [1], [2] right after the sentence they support. "
+        "Do not append a 'Sources:' section. If the answer is not in the context, say you do not know."
+    )
+
+    messages = [{"role": "system", "content": sys_prompt}]
+    if chat_history:
+        messages.extend(chat_history)
+    prompt = f"Context:\n{context}\n\nQuestion: {question}\nAnswer:"
+    messages.append({"role": "user", "content": prompt})
+
+    model_tag = resolve_model(model)
+
+    try:
+        r = requests.post(
+            f"{OLLAMA_URL}/api/chat",
+            json={
+                "model": model_tag,
+                "messages": messages,
+                "stream": False,
+                "options": {"temperature": 0.2, "num_predict": 300},
+                "keep_alive": OLLAMA_KEEP_ALIVE
+            },
+            timeout=120
+        )
+    except Exception as e:
+        raise HTTPException(status_code=502, detail=f"Ollama connection error: {e}")
+
+    if r.status_code != 200:
+        body = r.text
+        if len(body) > 800:
+            body = body[:800] + "...(truncated)"
+        raise HTTPException(status_code=502, detail=f"Ollama status {r.status_code}: {body}")
+
+    try:
+        j = r.json()
+    except Exception:
+        body = r.text
+        if len(body) > 800:
+            body = body[:800] + "...(truncated)"
+        raise HTTPException(status_code=502, detail=f"Ollama returned non-JSON: {body}")
+
+    msg = j.get("message") or {}
+    content = msg.get("content") or j.get("content") or ""
+
+    if not content.strip() and hits:
+        top = (hits[0].get("text") or "").strip()
+        if top and len(top) <= 200 and any(p in ql for p in ["what does", "say", "content", "quote", "exact text"]):
+            return f'It says: "{{ " ".join(top.split()) }}" [1].'
+
+    return dehedge(content)
